requirements.txt
Flask
pyTelegramBotAPI
rich
python-box
requests